backend/requirements.txt
asgiref==3.7.2
Django==4.2.8
sqlparse==0.4.4
typing_extensions==4.8.0
django-debug-toolbar==4.2.0
djangorestframework~=3.14.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
pillow==10.1.0
django-elasticsearch-dsl-drf==0.22.5